More fixes and improvements on autoencoder classes.
